+++ lib/init_video_player_media_kit/init_video_player_media_kit_ffi.dart
@@ -17,23 +17,25 @@ import 'package:video_player_media_kit/video_player_media_kit_platform/video_pla
 /// - `iosUseMediaKit`: A boolean value indicating whether to register the plugin on iOS.
 /// - `androidUseMediaKit`: A boolean value indicating whether to register the plugin on Android.
 /// - `logLevel`: A `LogLevel` value indicating the desired log level.
+/// - `throwErrors`: when playing errors happens, throw error (can be disabled incase of false positives)
 void initVideoPlayerMediaKitIfNeeded(
     {bool iosUseMediaKit = false,
     bool androidUseMediaKit = false,
-    MPVLogLevel logLevel = MPVLogLevel.warn}) {
+    MPVLogLevel logLevel = MPVLogLevel.warn,
+    bool throwErrors = true}) {
   if ((UniversalPlatform.isWindows ||
           UniversalPlatform.isLinux ||
           UniversalPlatform.isMacOS) &&
       !UniversalPlatform.isWeb) {
     MediaKit.ensureInitialized();
-    VideoPlayerMediaKit.registerWith(logLevel: logLevel);
+    VideoPlayerMediaKit.registerWith(logLevel: logLevel,throwErrors: throwErrors);
   }
   if (UniversalPlatform.isIOS && iosUseMediaKit) {
     MediaKit.ensureInitialized();
-    VideoPlayerMediaKit.registerWith(logLevel: logLevel);
+    VideoPlayerMediaKit.registerWith(logLevel: logLevel,throwErrors: throwErrors);
   }
   if (UniversalPlatform.isAndroid && androidUseMediaKit) {
     MediaKit.ensureInitialized();
-    VideoPlayerMediaKit.registerWith(logLevel: logLevel);
+    VideoPlayerMediaKit.registerWith(logLevel: logLevel,throwErrors: throwErrors);
   }
 }